backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
supabase>=2.0.0
google-genai>=1.0.0
slowapi>=0.1.9
python-multipart>=0.0.9
python-dotenv>=1.0.0
pydantic-settings>=2.2.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
